+++ django_spire/seeding/field/custom.py
@@ -1,3 +1,4 @@
+from django.db.models import QuerySet
 from django.utils import timezone
 from faker import Faker
 
@@ -9,8 +10,7 @@ class CustomFieldSeeder(BaseFieldSeeder):
     keyword = FieldSeederTypesEnum.CUSTOM
 
     def in_order(self, values: list, index: int) -> any:
-        if index >= len(values):
-            raise IndexError("Index exceeds the list length in 'in_order'")
+        index = index % len(values) # Index loops back on itself
         return values[index]
 
     def date_time_between(self, start_date: str, end_date: str):
@@ -19,11 +19,24 @@ class CustomFieldSeeder(BaseFieldSeeder):
         aware_dt = timezone.make_aware(naive_dt)
         return aware_dt
 
+    def fk_random(self, model_class, ids: list[int]):
+        faker = Faker()
+        return faker.random_element(elements=ids)
+
+    def fk_in_order(self, model_class, index: int, ids: list[int]):
+        """Takes a queryset, calls it then returns a random id"""
+        return self.in_order(values=ids, index=index)
+
     def seed(self, manager, count) -> list[dict]:
         data = []
+
+        foreign_keys = {}
+
         for i in range(count):
             row = {}
+
             for field_name, config in self.seeder_fields.items():
+
                 method_name = config[1] if len(config) > 1 else field_name
                 kwargs = config[2] if len(config) > 2 else {}
                 method = getattr(self, method_name, None)
@@ -31,8 +44,16 @@ class CustomFieldSeeder(BaseFieldSeeder):
                 if not callable(method):
                     raise ValueError(f"Custom method '{method_name}' not found for field '{field_name}'")
 
-                if method_name == "in_order":
+                if method_name in ['in_order', 'fk_in_order']:
                     kwargs["index"] = i
+
+                if method_name in ['fk_random', 'fk_in_order']:
+                    model_class = kwargs["model_class"]
+
+                    if model_class not in foreign_keys.keys():
+                        foreign_keys[model_class] = model_class.objects.all().values_list('id', flat=True)
+
+                    kwargs['ids'] = foreign_keys[model_class]
 
                 row[field_name] = method(**kwargs)