+++ docs/changelog/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.4.1
+
+### Changes
+#### Django Model Seeder
+- Custom methods to seed foreign keys to solve class variables initiation issues.
+- In order custom method loops back on itself to keep seeding values.
+- Django LLM Seeder processes to futures when the count is over 25.
+- Field Override checks seeder class when calling methods.
+
+
 ## v0.4.0
 
 ### Features